Generalised item reading. Games now use the same code to read, while retaining their own pointer paths. Improved readability of pointer path in PTDE. Refactored the item reading to read 1 big buffer from the games memory, then to read from this buffer, rather than doing individual reads. Speeds the code up by at least a factor 10.
