+++ src/DarkSoulsMemory/Internal/DarkSoulsRemastered/DarkSoulsRemastered.cs
@@ -182,14 +182,10 @@ namespace DarkSoulsMemory.Internal.DarkSoulsRemastered
 
         public List<Item> GetCurrentInventoryItems()
         {
-            //Path: GameDataMan->hostPlayerGameData->equipItemData->equipGameData->equipInventoryData->equipInventoryDataSub
-
-            var items = new List<Item>();
-
+            //Path: GameDataMan->hostPlayerGameData->equipGameData->equipInventoryData->equipInventoryDataSub
+            
             var hostPlayerGameData = (IntPtr)ReadInt32(_gameDataMan + 16);
-            var equipData = hostPlayerGameData + 0x280;
-            var equipItemData = equipData + 416;
-            var equipGameData = (IntPtr)ReadInt32(equipItemData + 16);
+            var equipGameData = hostPlayerGameData + 0x280; //640
             var equipInventoryData = equipGameData + 288;
             var equipInventoryDataSub = equipInventoryData + 16;
 
@@ -198,113 +194,15 @@ namespace DarkSoulsMemory.Internal.DarkSoulsRemastered
             var keyCount = ReadInt32(equipInventoryDataSub + 52);
 
             //Struct has 2 lists, list 1 seems to be a subset of list 2, the lists start at the same address..
-
+            //I think the first list only contains keys. The "master" list contains both.
             var itemList2Len = ReadInt32(equipInventoryDataSub);
-            var itemList1Len = ReadInt32(equipInventoryDataSub + 16);
+            //var itemList1Len = ReadInt32(equipInventoryDataSub + 16);
 
-            var itemList1 = (IntPtr)ReadInt32(equipInventoryDataSub + 32);
+            //var itemList1 = (IntPtr)ReadInt32(equipInventoryDataSub + 32);
             var itemList2 = (IntPtr)ReadInt32(equipInventoryDataSub + 40);
 
-            for (var i = 0; i < itemList2Len; i++)
-            {
-                //size of NS::FRPG_EquipInventoryDataItem is 28 or 0x1c
-                var addr = itemList2 + (i * 0x1c);
-                
-                var cat = ReadByte(addr + 3);
-                var item = ReadInt32(addr + 4);
-                var quantity = ReadInt32(addr + 8);//Is this right? gota test.
-
-
-                if (item != -1)
-                {
-                    var categories = new List<ItemCategory>();
-                    switch (cat.ToHex())
-                    {
-                        case 0:
-                            categories.Add(ItemCategory.MeleeWeapons);
-                            categories.Add(ItemCategory.RangedWeapons);
-                            categories.Add(ItemCategory.Ammo);
-                            categories.Add(ItemCategory.Shields);
-                            categories.Add(ItemCategory.SpellTools);
-                            break;
-
-                        case 1:
-                            categories.Add(ItemCategory.Armor);
-                            break;
-
-                        case 2:
-                            categories.Add(ItemCategory.Rings);
-                            break;
-
-                        case 4:
-                            categories.Add(ItemCategory.Consumables);
-                            categories.Add(ItemCategory.Key);
-                            categories.Add(ItemCategory.Spells);
-                            categories.Add(ItemCategory.UpgradeMaterials);
-                            categories.Add(ItemCategory.UsableItems);
-                            break;
-                    }
-
-                    //Decode item
-                    int id = 0;
-                    ItemInfusion infusion = ItemInfusion.Normal;
-                    int level = 0;
-
-                    //if 4 or less digits -> non-upgradable item.
-                    if (categories.Contains(ItemCategory.Consumables) && item >= 200 && item <= 215 && !items.Any(j => j.Type == ItemType.EstusFlask))
-                    {
-                        var estus = Item.AllItems.First(j => j.Type == ItemType.EstusFlask);
-                        var instance = new Item(estus.Name, estus.Id, estus.Type, estus.Category, estus.StackLimit, estus.Upgrade);
-
-                        //Item ID is both the item + reinforcement. Level field does not change in the games memory for the estus flask.
-                        //Goes like this:
-                        //200 == empty level 0
-                        //201 == full level 0
-                        //202 == empty level 1
-                        //203 == full level 1
-                        //203 == empty level 2
-                        //204 == full level 2
-                        //etc
-
-                        //If the flask is not empty, the amount of charges is stored in the quantity field.
-                        //If the ID - 200 is an even number, the flask is empty. For this case we can even ignore the 200 and just check the ID
-
-                        instance.Quantity = item % 2 == 0 ? 0 : quantity;
-
-                        //Calculating the upgrade level
-                        instance.UpgradeLevel = (item - 200) / 2;
-
-                        instance.Infusion = infusion;
-                        items.Add(instance);
-                        continue;
-                    }
-                    else if (item < 10000)
-                    {
-                        id = item;
-                    }
-                    else
-                    {
-                        //Separate digits
-                        int one = item % 10;
-                        int ten = (item / 10) % 10;
-                        int hundred = (item / 100) % 10;
-
-                        id = item - (one + (10 * ten) + (100 * hundred));
-                        infusion = (ItemInfusion)hundred;
-                        level = one + (10 * ten);
-                    }
-
-                    var lookupItem = Item.AllItems.FirstOrDefault(j => categories.Contains(j.Category) && j.Id == id);
-                    if (lookupItem != null)
-                    {
-                        var instance = new Item(lookupItem.Name, lookupItem.Id, lookupItem.Type, lookupItem.Category, lookupItem.StackLimit, lookupItem.Upgrade);
-                        instance.Quantity = quantity;
-                        instance.Infusion = infusion;
-                        instance.UpgradeLevel = level;
-                        items.Add(instance);
-                    }
-                }
-            }
+            var bytes = ReadBytes(itemList2, itemList2Len * 0x1c);
+            var items = ItemReader.GetCurrentInventoryItems(bytes, itemList2Len, itemCount, keyCount);
             
             return items;
         }
